refactor: remove unused funcs, comments, and move some code to the other file
